+++ CHANGES.md
@@ -5,6 +5,7 @@
 ### Bug fixes
 - When using the `PKG_CONFIG_PATH` specified by mopack, add it to any existing
   `PKG_CONFIG_PATH` from the environment
+- Don't explicitly include default `#include` paths on CC-like compilers
 
 ---